Implement a more perfomrant Box
I mentioned this in the code for Box::hit() and it's marked in ticket
 #3 (on GitLab).  I was thinking it would be more performant if we
didn't use the *Rect objects to construct a Box.  Instead, we use the
min & max point.  Then we stuffed the *Rect::hit() function's code into
our Box::hit().

The speedup on `book2::final_secne` wasn't as much as I liked it
(acutly faster), but the one on the cornel box scenes is really good.
E.g. on my machine, It's rendering ~20% faster.  On top of that, it no
longers requires an RNG to do `Box` constrution.  As my previous method
used BVHnodes for speedup.

A new flag was acced to the CMake file and some other misc. Cleanups
where made to the `Rectangles.cpp` file (removing dead TODOs).