Make the "Deep Copy Per Thread" a runtime flag
Makes it more flexible for testing.  Doesn't make any sense why it was
hard compiled in if it could be a runtime thing.

I noticed that on one of my laptops it helped multi-threaded performance
quite well, but on a (slightly older) one it actually rendered a little
slower.  I think this warrants some more investigation.