Add in a new BVHNode implementation that more performant
It's still veyr similar to that of the Books', except that it doesn't
acutally construct a tree using pointets and child nodes.  Instead, it
makes a vector of AABB's (that also contain object indicies), that then
are organized into the list and can be traversed like a tree.

The benefit is that this uses less memory and is faster.