Small experiment to make PDFs more efficient
I'm getting about 15% faster renders with this.  The idea is to
eliminate the usage of reference counted pointers, as well as
dynamically allocated memory creation/usage during the rendering process
(which is what the book does).

The trick here is to instead use a variant type of all the `IPDF` child
classes.  This puts the PDF on the stack instead of the heap.