Have the "earthmap.jpg" asset be embedded
It's the only image texture that's actually used, so it's much simplier
to embed its data in the program.  This is done because I don't want to
fiddle with distributing runtime assets for the iOS and Android port.

The `ImageTexture` constructor is now different from the book at two
"static constructor" methods are used instead.  One to load images from
a file, and another to load them from a memory buffer.

README has been updated to reflect this change.