Add in a GUI frontend using Qt 6 (Desktop & Mobile)
(Note that this is squashed from about 80 commits from a working branche
called `qt_ui`, which was worked on from around Nov. 1st 2021 till
January 17th 2022).

This commit adds in the "Qt UI", which acts as a frontend to all of the
command line options you'd normally see when trying to run the
PSRayTracing binary.  This was done primarily so I could test the
rendering performance of the project on mobile/tablet devices (i.e.
ARM), but then it started to grow into a much larger project in trying
to deliver a nice experieince.  Possibly serving as a sample app for
Qt/Qml development.

Features:
- A "Pan Zoom" image viewer for renders (still has some quirks)
  - Double tap to enter "detail view"
- Scene select dropdown (no loger have to type everything)
- Button to set to your device resolution
- Button to use a recommended thread count (i.e. the max concurrent
  threads supported on the device)

Development started with Qt 6.2.1, is now at Qt 6.2.2.  Along the way, I
found some bugs (reported them), and soon expect this to be bumped up to
.3 in the near future.

Devices tested on:
- Windows Desktop
- Windows 2-in-1
- Linux (Ubunutu)
- macOS (M1 & Intel)
- Android (Kindle Fire) Tablet
- Android Phone
- iPad (pro)

In the iOS simulator for iPhone (13), the UI didn't work quite
correctly, but at this point, I'd like to just release the initial
version.

At the moment of writing, I am waiting for Google to approve the app in
their Play store.  I would like to put this in the Apple App store in
the future.  So if you want to run the GUI frontend on those devices,
you'll need to build it yourself there.

---

Other changes:
- Need to turn of pedantic errors (and make them warnings instead)

---

Notes from (now squashed) commits:

=
  Added an "About Page"

  - Refactored new RenderView
  - Better scaling/sizing for high DPI displays
  - Have a resuable scrolling pane widget w/ scrollbar
    - Only goes up and down
    - Feels a little hacky in its implementation
  - Back button works for About Page
  - Added an "About Message"
    - Credited some constributors to the project
  - Defined what a "small width" is for the theme

=

  Have the Render view be responsive and work on different screens well

  - Refactoring/restructuring
    - anchorbased insteed of a row/column
  - Uses state framework

=

  Make a PanZoom control for the Renders

  It's not perfect (e.g. doesn't center on the zoom point), but I think
  it's good for a first past.  Seems to work adequate on Android (phone &
  tablet), well on Linux touchscreen/desktop, and I assume it's good on
  Mac and iPad (will test).  You pinch/scroll to zoom.  Swipe or
  click-n-hold to pan. etc.

  To enable the mode, you'll need a completed render and have to either
  double tap or double click on it.  To get out of the mode, double
  tap/click again.

  Working on double-tap handler for rnedered iamge