Have the `master` branch support Book 3
Before anyone would have to checkout the `book3` branch separately if
they wanted to see the renders from that book.  If I had a change in
`master`, it was honestly a pain to also update that branch.

Instead, it's MUCH nicer to support all three books in the same branch.
I was worried about the complexity of this when first emabarking on book
3.  But now I realized it wasn't that difficult. I did have to add some
stuff on the books' original code (e.g. see the additions to the
`IMaterial` interface).  The important one being `RenderMethod`, so we
can delinatate if we are using Book 1 & 2 style renders, or Book 3.

I was a little suprised, but there seem to be no noticable performance
hits.  Which is great. On top of this, the GUI app can now show all
scenes at the same time without having to rebuild the software.  I think
an app store update is in order.

Note that there were also some code naming cleanups.

Fixes #11