Cleanup the `get_in_unit_sphere()` RNG code
Similar to trying to fixup the `get_in_unit_disk()` code a commit or two
earlier, I was trying to find a loop-less/branch-less method.  I found
this one, but turns out the book's method is still better.  I'm leaving
in the other code for anyone else to look at.