Remove CMake line specifying iPhone/iPad as targets
This was QTBUG-100833 and was resolved in Qt 6.3.1.  CMake was
originally only generating xcode targets for iPhone, when we also needed
iPad as well.  We had to have that line in there, but now it's specified
by default with Qt.  So we don't need it anymore and can remove it.

Fixes #41